Generate ROC curves to evaluate TMB performance in immunotherapy responder classification

This function takes in input a data frame with a vector of TMB values and a factor named ClinicalResponse with only these two levels allowed: 'responder' and 'nonresponder'. It gives in output a ROC curve describing TMB performance in responders and nonresponders classification on a training set (random 75 of input data), the best value of TMB cutoff for classification in this training dataset, a confusion matrix describing the classification performance of this TMB cutoff in a training dataset (25 describing TMB performance in responders and nonresponders classification in a training set.

generateROC(dataset, method = "Youden")

Arguments

dataset
a data.frame containing a numeric vector of TMB values and a factor with name 'ClinicalResponse' and levels 'responder' and 'nonresponder' and a column with TMB values named "TMB_per_Mb".

method
character string indicating the method to identify the best TMB cutoff: i.e. Youden, MaxSe, MaxSpSe. For more details see also optimal.cutpoints

Value

Returns in output the best TMB cutoff for classification, AUC with 95 values based on the model built on training data. It also returns the confusion matrix of classification on test data based on the identified TMB cutoff and a the ROC curve on train and test data.

Author

Laura Fancello