Read design file

Read the file describing the design of the sequencing experiment (i.e. the targeted regions in the sequencing experiment). It can be either a BED file containing genomic coordinates of the targeted genes or a TXT file containing symbols for targeted genes.

readDesign(filename, assembly, ids, name = NULL)

Arguments

filename
name of file describing the sequencing design. It requires ".bed" extension for files containing the genomic coordinates of the targeted regions or ".txt" extension for files containing the gene symbol of targeted genes. BED files must be tab-delimited, with no track description line and it must report at least CHR START and END information.

assembly
human genome assembly: hg19 or hg38

ids
type of gene identifiers, if .txt file with gene list is provided: hgnc_symbol or entrezgene_id.

name
string indicating the name of the design. if not specified, the filename without extension will be used instead.

Value

data.frame with genomic coordinates or a character vector with symbols of targeted genes

Author

Laura Fancello

Examples

#Read in input the panel sequencing design (either a BED file with genomic
#coordinates of the targeted genes or TXT file with targeted gene names).
design_FromTxt <- readDesign(filename = system.file("extdata",
                                 "ExamplePanel_GeneIDs.txt",
                                 package = "TMBleR",
                                 mustWork = TRUE),
                     assembly = "hg19",
                     ids = "entrezgene_id")

#> Cache found

design_FromBed <- readDesign(filename = system.file("extdata",
                                "ExamplePaneldesign.bed",
                                 package = "TMBleR",
                                 mustWork = TRUE),
                              assembly = "hg19")